Moved GlobalExportWriter into its own file
