+++ specs/.eslintrc
@@ -1,5 +1,8 @@
 {
   "env": {
     "mocha": true
+  },
+  "ecmaFeatures": {
+    "modules": true
   }
 }
